+++ air/tests/test_module/negative_tests/execution_step.rs
@@ -14,18 +14,22 @@
  * limitations under the License.
  */
 
+use air::no_error_last_error_object;
 use air::unsupported_map_key_type;
 use air::CatchableError;
 use air::LambdaError;
 
+use air_test_framework::AirScriptExecutor;
 use air_test_utils::prelude::*;
 
 #[test]
 fn local_service_error() {
     let client_peer_id = "some_peer_id";
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (call "{client_peer_id}" ("m" "f") ["arg1"] other)
-    "#);
+    "#
+    );
 
     let mut client_vm = create_avm(echo_call_service(), client_peer_id);
     let result = client_vm
@@ -51,51 +55,8 @@ fn local_service_error() {
         )
         .unwrap();
 
-    let err_msg = std::rc::Rc::new(f!("\"{err_msg}\""));
+    let err_msg = std::rc::Rc::new(format!("\"{err_msg}\""));
     let expected_error = CatchableError::LocalServiceError(10000, err_msg);
-    assert!(check_error(&result, expected_error));
-}
-
-#[test]
-fn variable_not_found_ap_scalar() {
-    let vm_2_peer_id = "vm_2_peer_id";
-    let var_name = "scalar_1".to_string();
-    let script = f!(r#"
-        (par
-            (match 0 1
-                (call "vm_1_peer_id" ("m" "f") ["scalar_1_result"] {var_name})
-            )
-            (ap {var_name} scalar_2)
-        )
-        "#);
-    let mut vm_2 = create_avm(echo_call_service(), vm_2_peer_id);
-    let result = vm_2
-        .call_single(&script, "", "", vm_2_peer_id, 0, 0, None, <_>::default(), "")
-        .unwrap();
-    let expected_error = CatchableError::VariableNotFound(var_name);
-    assert!(check_error(&result, expected_error));
-}
-
-#[test]
-fn variable_not_found_ap_canon_stream() {
-    let vm_2_peer_id = "vm_2_peer_id";
-    let var_name = "#canon".to_string();
-    let script = f!(r#"
-        (par
-            (match 0 1
-                (seq
-                    (call "vm_1_peer_id" ("m" "f") ["scalar_1_result"] $stream)
-                    (canon "vm_1_peer_id" $stream {var_name})
-                )
-            )
-            (ap {var_name} scalar_2)
-        )
-        "#);
-    let mut vm_2 = create_avm(echo_call_service(), vm_2_peer_id);
-    let result = vm_2
-        .call_single(&script, "", "", vm_2_peer_id, 0, 0, None, <_>::default(), "")
-        .unwrap();
-    let expected_error = CatchableError::VariableNotFound(var_name);
     assert!(check_error(&result, expected_error));
 }
 
@@ -103,7 +64,8 @@ fn variable_not_found_ap_canon_stream() {
 fn fold_iterates_over_non_array_scalar_iterable() {
     let vm_2_peer_id = "vm_2_peer_id";
     let var_name = "outcome".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_2_peer_id}" ("m" "f") [] {var_name})
             (fold {var_name} i
@@ -113,7 +75,8 @@ fn fold_iterates_over_non_array_scalar_iterable() {
                 )
             )
         )
-        "#);
+        "#
+    );
     let unsupported_jvalue = json!({"attr": 42, });
     let mut vm_2 = create_avm(set_variable_call_service(unsupported_jvalue.clone()), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
@@ -126,7 +89,8 @@ fn fold_iterates_over_non_array_scalar_lambda_iterable() {
     let vm_2_peer_id = "vm_2_peer_id";
     let var_name = "outcome".to_string();
     let scalar_int = 42;
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_2_peer_id}" ("m" "f") [] {var_name})
             (fold {var_name}.$.int i
@@ -136,7 +100,8 @@ fn fold_iterates_over_non_array_scalar_lambda_iterable() {
                 )
             )
         )
-        "#);
+        "#
+    );
     let unsupported_jvalue = json!({"int": scalar_int, });
     let mut vm_2 = create_avm(set_variable_call_service(unsupported_jvalue.clone()), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
@@ -149,12 +114,14 @@ fn non_string_value_in_triplet_resolution() {
     let vm_2_peer_id = "vm_2_peer_id";
     let scalar_int = 42;
     let var_name = "var_name".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (ap {scalar_int} {var_name})
             (call {var_name} ("" "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(unit_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -163,7 +130,8 @@ fn non_string_value_in_triplet_resolution() {
     };
     assert!(check_error(&result, expected_error));
 
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (call "{vm_2_peer_id}" ("m" "f") [] {var_name})
@@ -171,7 +139,8 @@ fn non_string_value_in_triplet_resolution() {
             )
             (call {var_name}.$.int ("" "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(set_variable_call_service(json!({"int": scalar_int, })), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -181,7 +150,8 @@ fn non_string_value_in_triplet_resolution() {
     assert!(check_error(&result, expected_error));
 
     let var_name = "#canon".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (ap {scalar_int} $stream)
@@ -189,7 +159,8 @@ fn non_string_value_in_triplet_resolution() {
             )
             (call #canon.$.[0] ("" "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(echo_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -204,12 +175,14 @@ fn non_string_value_in_triplet_resolution_module_name() {
     let vm_2_peer_id = "vm_2_peer_id";
     let scalar_int = 42;
     let var_name = "var_name".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (ap {scalar_int} {var_name})
             (call "{vm_2_peer_id}" ({var_name} "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(unit_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -218,7 +191,8 @@ fn non_string_value_in_triplet_resolution_module_name() {
     };
 
     assert!(check_error(&result, expected_error));
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (call "{vm_2_peer_id}" ("m" "f") [] {var_name})
@@ -226,7 +200,8 @@ fn non_string_value_in_triplet_resolution_module_name() {
             )
             (call "{vm_2_peer_id}" ({var_name}.$.int "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(set_variable_call_service(json!({"int": scalar_int, })), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -236,7 +211,8 @@ fn non_string_value_in_triplet_resolution_module_name() {
     assert!(check_error(&result, expected_error));
 
     let var_name = "#canon".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (ap {scalar_int} $stream)
@@ -244,7 +220,8 @@ fn non_string_value_in_triplet_resolution_module_name() {
             )
             (call "{vm_2_peer_id}" (#canon.$.[0] "") [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(echo_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -259,12 +236,14 @@ fn non_string_value_in_triplet_resolution_function_name() {
     let vm_2_peer_id = "vm_2_peer_id";
     let scalar_int = 42;
     let var_name = "var_name".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (ap {scalar_int} {var_name})
             (call "{vm_2_peer_id}" ("" {var_name}) [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(unit_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -273,7 +252,8 @@ fn non_string_value_in_triplet_resolution_function_name() {
     };
 
     assert!(check_error(&result, expected_error));
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (call "{vm_2_peer_id}" ("m" "f") [] {var_name})
@@ -281,7 +261,8 @@ fn non_string_value_in_triplet_resolution_function_name() {
             )
             (call "{vm_2_peer_id}" ("" {var_name}.$.int) [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(set_variable_call_service(json!({"int": scalar_int, })), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -291,7 +272,8 @@ fn non_string_value_in_triplet_resolution_function_name() {
     assert!(check_error(&result, expected_error));
 
     let var_name = "#canon".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (seq
                 (ap {scalar_int} $stream)
@@ -299,7 +281,8 @@ fn non_string_value_in_triplet_resolution_function_name() {
             )
             (call "{vm_2_peer_id}" ("" #canon.$.[0]) [] outcome)
         )
-        "#);
+        "#
+    );
     let mut vm_2 = create_avm(echo_call_service(), vm_2_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::NonStringValueInTripletResolution {
@@ -313,12 +296,14 @@ fn non_string_value_in_triplet_resolution_function_name() {
 fn value_not_contain_such_array_idx_ap() {
     let vm_1_peer_id = "vm_1_peer_id";
     let idx = 42;
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_1_peer_id}" ("m" "f") [] result)
             (ap result.$.[{idx}].atrib outcome)
         )
-        "#);
+        "#
+    );
     let wrong_jvalue = json!(["a", "b"]);
     let mut vm_2 = create_avm(set_variable_call_service(wrong_jvalue.clone()), vm_1_peer_id);
     let result = vm_2.call(&script, "", "", <_>::default()).unwrap();
@@ -337,12 +322,14 @@ fn field_accessor_applied_to_stream() {
     let arg = json!({"a": 1,});
     let mut peer_vm_1 = create_avm(set_variable_call_service(arg), vm_peer_id_1);
     let field_name = "int".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (canon "{vm_peer_id_1}" $stream #canon_stream)
             (call "{vm_peer_id_1}" ("m" "f") [#canon_stream.$.{field_name}] output)
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error =
@@ -357,12 +344,14 @@ fn array_accessor_not_match_value() {
     let arg = json!({"a": 1,});
     let mut peer_vm_1 = create_avm(set_variable_call_service(arg.clone()), vm_peer_id_1);
     let idx = 42;
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] nonarray)
             (call "{vm_peer_id_1}" ("m" "f") [nonarray.$.[{idx}]] result)
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error =
@@ -376,12 +365,14 @@ fn value_not_contain_such_array_idx_call_arg_lambda() {
     let arg = json!([0, 1]);
     let mut peer_vm_1 = create_avm(set_variable_call_service(arg.clone()), vm_peer_id_1);
     let idx = 42;
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] array)
             (call "{vm_peer_id_1}" ("m" "f") [array.$.[{idx}]] result)
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error =
@@ -395,12 +386,14 @@ fn value_not_contain_such_field_call_arg_lambda() {
     let arg = json!({"a": 1,});
     let mut peer_vm_1 = create_avm(set_variable_call_service(arg.clone()), vm_peer_id_1);
     let field_name = "b".to_string();
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] obj)
             (call "{vm_peer_id_1}" ("m" "f") [obj.$.{field_name}] output)
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error =
@@ -414,12 +407,14 @@ fn field_accesssor_not_match_value_call_arg_lambda() {
     let arg = json!([0, 1]);
     let mut peer_vm_1 = create_avm(set_variable_call_service(arg.clone()), vm_peer_id_1);
 
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] obj)
             (call "{vm_peer_id_1}" ("m" "f") [obj.$.b] output)
         )
-        "#);
+        "#
+    );
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error = air::CatchableError::LambdaApplierError(air::LambdaError::FieldAccessorNotMatchValue {
         value: arg,
@@ -438,7 +433,8 @@ fn index_access_not_u32_i64() {
     let arg = json!([0, 1]);
     let mut peer_vm_2 = create_avm(set_variable_call_service(arg.clone()), vm_peer_id_2);
 
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] idx)
             (seq
@@ -446,7 +442,8 @@ fn index_access_not_u32_i64() {
                 (call "{vm_peer_id_2}" ("m" "f") [array.$.[idx]] result)
             )
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let result = peer_vm_2.call(script.clone(), "", result.data, <_>::default()).unwrap();
@@ -461,7 +458,8 @@ fn scalar_accessor_has_invalid_type_ap() {
     let obj_arg = json!({"a": 1,});
     let mut peer_vm_1 = create_avm(set_variable_call_service(obj_arg.clone()), vm_peer_id_1);
 
-    let script = f!(r#"
+    let script = format!(
+        r#"
         (seq
             (call "{vm_peer_id_1}" ("m" "f") [] array_obj)
             (seq
@@ -469,7 +467,8 @@ fn scalar_accessor_has_invalid_type_ap() {
                 (ap array.$.[array_obj] outcome)
             )
         )
-        "#);
+        "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error = air::CatchableError::LambdaApplierError(air::LambdaError::ScalarAccessorHasInvalidType {
@@ -492,7 +491,8 @@ fn stream_accessor_has_invalid_type() {
     let obj_arg = json!({"a": 1,});
     let mut peer_vm_1 = create_avm(set_variable_call_service(obj_arg.clone()), vm_peer_id_1);
 
-    let script = f!(r#"
+    let script = format!(
+        r#"
     (seq
         (call "{vm_peer_id_1}" ("m" "f") [] array_obj)
         (seq
@@ -503,7 +503,8 @@ fn stream_accessor_has_invalid_type() {
             )
         )
     )
-    "#);
+    "#
+    );
 
     let result = peer_vm_1.call(script.clone(), "", "", <_>::default()).unwrap();
     let expected_error = air::CatchableError::LambdaApplierError(air::LambdaError::StreamAccessorHasInvalidType {
@@ -525,7 +526,8 @@ fn canon_stream_not_have_enough_values_call_arg() {
     let local_peer_id = "local_peer_id";
     let mut local_vm = create_avm(echo_call_service(), local_peer_id);
 
-    let join_stream_script = f!(r#"
+    let join_stream_script = format!(
+        r#"
     (seq
         (seq
             (call "{local_peer_id}" ("" "") [[]] nodes)
@@ -540,7 +542,8 @@ fn canon_stream_not_have_enough_values_call_arg() {
             (canon "{local_peer_id}" $ns #ns)
             (call "{local_peer_id}" ("" "") [#ns.$.[0] #ns.$.[1] #ns])
         )
-     )"#);
+     )"#
+    );
 
     let result = local_vm.call(&join_stream_script, "", "", <_>::default()).unwrap();
     let actual_trace = trace_from_result(&result);
@@ -555,9 +558,11 @@ fn unsupported_map_keytype_float() {
     let mut local_vm = create_avm(unit_call_service(), "local_peer_id");
 
     let map_name = "%map";
-    let join_stream_script = f!(r#"
+    let join_stream_script = format!(
+        r#"
         (ap (0.5 "serv1") %map)
-     "#);
+     "#
+    );
 
     let result = local_vm.call(&join_stream_script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::StreamMapError(unsupported_map_key_type(map_name));
@@ -571,14 +576,87 @@ fn unsupported_map_keytype() {
     let mut local_vm = create_avm(set_variable_call_service(obj_arg), local_peer_id);
 
     let map_name = "%map";
-    let join_stream_script = f!(r#"
+    let join_stream_script = format!(
+        r#"
     (seq
         (call "{local_peer_id}" ("" "") [] scalar)
         (ap (scalar.$.a "serv1") %map)
     )
-     "#);
+     "#
+    );
 
     let result = local_vm.call(&join_stream_script, "", "", <_>::default()).unwrap();
     let expected_error = CatchableError::StreamMapError(unsupported_map_key_type(map_name));
     assert!(check_error(&result, expected_error));
+}
+
+#[test]
+fn undefined_last_error_functor() {
+    let local_peer_id = "local_peer_id";
+    let script = format!(
+        r#"
+    (xor
+        (match 1 2
+            (null)
+        )
+        (call "local_peer_id" ("test" "error_code") [%last_error%.length] scalar)
+    )
+    "#
+    );
+
+    let executor = AirScriptExecutor::from_annotated(TestRunParameters::from_init_peer_id(local_peer_id), &script)
+        .expect("invalid test AIR script");
+    let result = executor.execute_all(local_peer_id).unwrap();
+
+    let expected_error = CatchableError::LengthFunctorAppliedToNotArray(no_error_last_error_object());
+    assert!(check_error(&result.last().unwrap(), expected_error));
+}
+
+#[test]
+fn undefined_last_error_instruction() {
+    let local_peer_id = "local_peer_id";
+    let script = format!(
+        r#"
+    (xor
+        (match 1 2
+            (null)
+        )
+        (call "local_peer_id" ("test" "instruction") [%last_error%.$.instruction] scalar)
+    )
+    "#
+    );
+
+    let executor = AirScriptExecutor::from_annotated(TestRunParameters::from_init_peer_id(local_peer_id), &script)
+        .expect("invalid test AIR script");
+    let result = executor.execute_all(local_peer_id).unwrap();
+
+    let expected_error = CatchableError::LambdaApplierError(LambdaError::ValueNotContainSuchField {
+        value: no_error_last_error_object(),
+        field_name: "instruction".to_string(),
+    });
+    assert!(check_error(&&result.last().unwrap(), expected_error));
+}
+
+#[test]
+fn undefined_last_error_peer_id() {
+    let local_peer_id = "local_peer_id";
+    let script = format!(
+        r#"
+    (xor
+        (match 1 2
+            (null)
+        )
+        (call "local_peer_id" ("test" "peer_id") [%last_error%.$.peer_id] scalar)
+    )
+    "#
+    );
+    let executor = AirScriptExecutor::from_annotated(TestRunParameters::from_init_peer_id(local_peer_id), &script)
+        .expect("invalid test AIR script");
+    let result = executor.execute_all(local_peer_id).unwrap();
+
+    let expected_error = CatchableError::LambdaApplierError(LambdaError::ValueNotContainSuchField {
+        value: no_error_last_error_object(),
+        field_name: "peer_id".to_string(),
+    });
+    assert!(check_error(&&result.last().unwrap(), expected_error));
 }