+++ crates/beautifier/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [0.2.0](https://github.com/fluencelabs/aquavm/compare/air-beautifier-v0.1.4...air-beautifier-v0.2.0) (2023-07-16)
+
+
+### ⚠ BREAKING CHANGES
+
+* **execution-engine:** stream map to scalar conversion using canon instruction [fixes VM-294] ([#610](https://github.com/fluencelabs/aquavm/issues/610))
+
+### Features
+
+* **execution-engine:** Stream Map to Scalar conversion using canon instruction [fixes VM-294] ([fcb4c9d](https://github.com/fluencelabs/aquavm/commit/fcb4c9dab43d82e87f1d6f8a83b72f6891d37bef))
+* **execution-engine:** stream map to scalar conversion using canon instruction [fixes VM-294] ([#610](https://github.com/fluencelabs/aquavm/issues/610)) ([fcb4c9d](https://github.com/fluencelabs/aquavm/commit/fcb4c9dab43d82e87f1d6f8a83b72f6891d37bef))
+
+
+### Bug Fixes
+
+* **air-beautifier:** print fold's last_instruction ([#630](https://github.com/fluencelabs/aquavm/issues/630)) ([adcf708](https://github.com/fluencelabs/aquavm/commit/adcf708c95baba1df319e438152a71607f52e92b))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * aquavm-air-parser bumped from 0.7.5 to 0.8.0
+
 ## [0.1.4](https://github.com/fluencelabs/aquavm/compare/air-beautifier-v0.1.3...air-beautifier-v0.1.4) (2023-06-22)