+++ tools/wasm/air-beautify-wasm/CHANGELOG.md
@@ -4,6 +4,10 @@
   * dependencies
     * air-beautifier bumped from 0.1.3 to 0.1.4
 
+* The following workspace dependencies were updated
+  * dependencies
+    * air-beautifier bumped from 0.1.4 to 0.2.0
+
 ## [0.3.0](https://github.com/fluencelabs/aquavm/compare/air-beautify-wasm-v0.2.1...air-beautify-wasm-v0.3.0) (2023-06-23)